[NFC] Restructure the passes module
